Moved evaluation functions and writing files outside of main.
